requirements.txt
fastapi
uvicorn
python-multipart
langchain
openai
chromadb
pymupdf
tiktoken
python-dotenv
pytesseract 
pillow
git-filter-repo 